# portal/source/IBMG.SCS.Portal.ApiClient/OpenAPI/kerridge.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "IBMG.SCS.KerridgeApi.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/customer/{customerAccountNumber}/spend-limit/{newSpendLimit}": {
      "get": {
        "tags": [
          "Spend"
        ],
        "summary": "New spend limit for a customer",
        "description": "Sets a new spend limit for a customer.",
        "operationId": "NewSpendLimit",
        "parameters": [
          {
            "name": "customerAccountNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newSpendLimit",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpendLimitResponseDtoListApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SpendLimitResponseDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "newSpendLimit": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "SpendLimitResponseDtoListApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpendLimitResponseDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "X-Branch-Code": {
        "type": "apiKey",
        "description": "Branch code header",
        "name": "X-Branch-Code",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    },
    {
      "X-Branch-Code": [ ]
    }
  ]
}